% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R
\name{analyse}
\alias{analyse}
\title{analyse Function}
\usage{
analyse(column, cn, counter, na_drop = T)
}
\arguments{
\item{column}{column name}

\item{cn}{counter}

\item{counter}{counter}

\item{na_drop}{Drop na? default is True}
}
\description{
This is a helper function
}
\keyword{analyse}
