% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{as.data.frame.obm_class}
\alias{as.data.frame.obm_class}
\title{as.data.frame class Function}
\usage{
\method{as.data.frame}{obm_class}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{\code{obm_class} S3 class object}

\item{row.names}{Optional row names argument (ignored).}

\item{optional}{Logical flag (ignored). Included for consistency with \code{as.data.frame}.}

\item{...}{Additional arguments (ignored).}
}
\value{
data.frame
}
\description{
Extract the underlying data frame from an \code{obm_class} S3 object.
}
\examples{
\dontrun{
as.data.frame(obm_data)
}
}
\keyword{as.data.frame}
