% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{as.obm_class}
\alias{as.obm_class}
\title{Internal function to create obm_class from data.frame}
\usage{
as.obm_class(x)
}
\value{
structured data
}
\description{
This internal function creates an obm_class
}
\keyword{internal}
