% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R
\name{create_table}
\alias{create_table}
\title{create_table Function}
\usage{
create_table(
  file = NULL,
  data = NULL,
  sep = ",",
  quote = "'",
  createtable = FALSE,
  project = FALSE,
  table = NULL,
  na_value = ""
)
}
\arguments{
\item{file}{csv file for processing}

\item{data}{instead of input file using a data.frame as input object}

\item{sep}{separator character in csv file}

\item{quote}{quote character of fields in csv file}

\item{createtable}{output as CREATE TABLE ... or ALTER TABLE ... ADD COLUMN ...}

\item{project}{prefix for table name}

\item{table}{the output sql table name}

\item{na_value}{Ommitted na values in type analises. na_value can be a vector or a single value as well. Default is empty the cell "".}
}
\description{
Create table structure from csv file
./create_table_from_csv.R --file foo.csv [--sep , --quote \' --createtable --project ... --table ... --na_value ...]
Default for quote is "
Default for sep is ,
Default for createtable is FALSE. If TRUE, SQL output will be CREATE TABLE... instead of ALTER TABLE... 
No default for project. If set, table name will be prefixed with this value
Default table is the basename of the csv file.
}
\keyword{create}
\keyword{sql}
\keyword{structure}
\keyword{table}
