% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obm_env.R
\name{get_obm}
\alias{get_obm}
\title{Retrieve a variable from the OBM session environment}
\usage{
get_obm(name, default = NULL)
}
\description{
Retrieve a variable from the OBM session environment
}
\keyword{internal}
