% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R, R/mark_import.R
\name{getExtension}
\alias{getExtension}
\title{getExtension Function}
\usage{
getExtension(file)

getExtension(file)
}
\arguments{
\item{file}{is a file name}
}
\description{
This is a helper function

This is a helper function
}
\keyword{extension}
\keyword{file}
\keyword{file-extension}
