% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_import.R
\name{mark_transform}
\alias{mark_transform}
\title{mark_transform Function}
\usage{
mark_transform(
  file = NULL,
  data = NULL,
  csv.sep = NULL,
  csv.quote = NULL,
  write_output_files = T,
  na_values = "",
  na_drop = F
)
}
\arguments{
\item{file}{A CSV or Excel file path}

\item{data}{Optional data.frame instead of input file}

\item{csv.sep}{CSV separator string. Default is autodetect}

\item{csv.quote}{CSV quote string. Default is autodetect}

\item{write_output_files}{Write all outputs to separated files}

\item{na_values}{Replace specified characters to NA, read.csv->na.strings}

\item{na_drop}{Skip error evaluation of lines where control-column value is empty}
}
\description{
Import a csv file and transform to Mark input file
./mark_transform.R --file foo.csv [--sep , --quote \' ... ]
}
\examples{
\dontrun{
mark_transform(file='maculinea_2021.07.19.csv')
}
}
