% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{obm_auth}
\alias{obm_auth}
\title{Log in to an OpenBioMaps server}
\usage{
obm_auth(username = "", password = "", verbose = FALSE, paranoid = TRUE)
}
\arguments{
\item{username}{Your OBM username (email)}

\item{password}{Your password}

\item{verbose}{print verbose feedback messages - default is FALSE}

\item{paranoid}{hide password while typing (on Linux) - default is TRUE}
}
\value{
true if successful and false if not
}
\description{
This function allows you to connect to and OBM server.
}
\examples{
\dontrun{
# default interactive authentication
obm_auth()

# authentication with username and password
obm_auth('foo@google.com','abc123')
}
}
\keyword{authentication}
\keyword{login}
