% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{obm_computation}
\alias{obm_computation}
\title{Computation Function}
\usage{
obm_computation(
  action = "",
  data_files = NULL,
  params = NULL,
  config_file = NULL
)
}
\arguments{
\item{action}{post, get-results, get-status}

\item{data_files}{list of files to sending}

\item{params}{computation control params}

\item{config_file}{is The computation_conf.yml file}
}
\value{
data frame or boolean
}
\description{
This function allows you to get data from an OpenBioMaps server.
}
\examples{
\donttest{
# Manage computational packs
results <- obm_computation('post','afile.R',params->c())
}
}
\keyword{computation}
