% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{obm_set}
\alias{obm_set}
\title{Set Function}
\usage{
obm_set(scope = "", condition = "")
}
\arguments{
\item{scope}{Which scope? e.g. set_join}

\item{condition}{A text condition based on column in your table}
}
\value{
list, json or boolean
}
\description{
Experimental function in APIv2!
This function allows you to set rules for obm_get.
}
\examples{
\donttest{
# automatically join tables
data <- obm_set('set_join',c('dead_animals','dead_animals_history'))
}
}
\keyword{set}
