% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{randtext}
\alias{randtext}
\title{Helper Function}
\usage{
randtext(n = 5000)
}
\arguments{
\item{n}{replication}
}
\value{
text
}
\description{
This function allows put data into a repozitorium.
}
\keyword{random}
\keyword{text}
