% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{repo_get}
\alias{repo_get}
\title{repo_get Function}
\usage{
repo_get(
  scope = NULL,
  token = OBM$token,
  pds_url = OBM$pds_url,
  data_table = OBM$project
)
}
\arguments{
\item{obm_class}{S3 class object}
}
\description{
This is repozitorium function
}
\keyword{as.data.frame}
