% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_import.R
\name{x_mark_transform}
\alias{x_mark_transform}
\title{x_mark_transform Function}
\usage{
x_mark_transform(
  data,
  id_col,
  date_col,
  species_name_col = NULL,
  sex_col = NULL,
  comment_col = NULL,
  write_output_files = T,
  na_drop = F
)
}
\arguments{
\item{data}{(data.frame)}

\item{id_col}{(character string)}

\item{date_col}{(character string)}

\item{species_name_col}{(character string or NULL)}

\item{sex_col}{(character string or NULL)}

\item{comment_col}{(character string or NULL)}

\item{write_output_files}{(T/F)}

\item{na_drop}{drop na? - skip error evaluation of lines where control-column value is empty}
}
\description{
This is a helper function
}
\keyword{Mark}
\keyword{error-check}
\keyword{transform}
