% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R
\name{analyse}
\alias{analyse}
\title{analyse Function}
\usage{
analyse(col, cn, counter, na.drop = T)
}
\arguments{
\item{cn}{counter}

\item{counter}{counter}

\item{column}{column name}

\item{na}{drop na?}
}
\description{
This is a helper function
}
\keyword{analyse}
