% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{as.data.frame.obm_class}
\alias{as.data.frame.obm_class}
\title{as.data.frame class Function}
\usage{
\method{as.data.frame}{obm_class}(x)
}
\arguments{
\item{obm_class}{S3 class object}
}
\description{
This class function extract data frame from obm_class S3 object
}
\examples{
as.data.frame(obm.data)
}
\keyword{as.data.frame}
