% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{as.obm_class}
\alias{as.obm_class}
\title{as. class Function}
\usage{
as.obm_class(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
object
}
\description{
This class function creates an obm_class
}
\examples{
\dontrun{
as.obm_class(DF)
}
}
\keyword{as}
\keyword{obm_class}
