% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{obm_auth}
\alias{obm_auth}
\title{Auth Function}
\usage{
obm_auth(
  username = "",
  password = "",
  scope = OBM$scope,
  client_id = OBM$client_id,
  url = OBM$token_url,
  verbose = F,
  paranoid = T
)
}
\arguments{
\item{username}{Your OBM username (email)}

\item{password}{Your password}

\item{scope}{vector for OAuth2 scopes}

\item{client_id}{Default is R}

\item{url}{OAuth2 token url obm_init() provide it}

\item{verbose}{print some messages}

\item{paranoid}{hide password while typing (on Linux)}
}
\description{
This function allows you to connect to and OBM server.
}
\examples{
\dontrun{
obm_auth()
token <- obm_auth('foo@google.com','abc123')
}
}
\keyword{auth}
