% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{OBM_auth}
\alias{OBM_auth}
\title{Auth Function}
\usage{
OBM_auth(username = "", password = "", scope = OBM$scope,
  client_id = OBM$client_id, url = OBM$token_url, verbose = F,
  paranoid = F)
}
\arguments{
\item{username}{Your OBM username (email)}

\item{password}{Your password}

\item{scope}{vector OAuth2 scopes scecified in the server DEFAULTS are: get_form_data get_form_list get_profile get_data get_history}

\item{client_id}{Default is R}

\item{url}{OAuth2 token url OBM_init() provide it}

\item{verbose}{print some messages}

\item{paranoid}{hide password while typing (on Linux)}
}
\description{
This function allows you to connect to and OBM server.
}
\examples{
OBM_auth()
token <- OBM_auth('foo@google.com','abc123')
}
\keyword{auth}
