% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{obm_computation}
\alias{obm_computation}
\title{Computation Function}
\usage{
obm_computation(
  action = "",
  token = OBM$token,
  url = OBM$pds_url,
  data_file = NULL,
  params = NULL
)
}
\arguments{
\item{action}{post, get-results, get-status}

\item{token}{obm_init() provide it}

\item{url}{obm_init() provide it}

\item{data_file}{data file sending for computation}

\item{params}{computation control params}
}
\description{
This function allows you to get data from an OpenBioMaps server.
}
\examples{
Manage computational packs
results <- obm_computation('post','afile.R',params->c())
}
\keyword{computation}
