% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{obm_get}
\alias{obm_get}
\title{Get Function}
\usage{
obm_get(
  scope = "",
  control_condition = NULL,
  condition = NULL,
  token = OBM$token,
  url = OBM$pds_url,
  table = OBM$project
)
}
\arguments{
\item{scope}{Which scope? e.g. get_data}

\item{control_condition}{An SQL control condition, e.g. limit=10:1}

\item{condition}{list - A condition based on column in your table, mostly key value pairs obm_get('get_data',condition = list(species = 'Parus palustris'))}

\item{token}{obm_init() provide it}

\item{url}{obm_init() provide it}

\item{table}{optional table from project}
}
\description{
This function allows you to get data from an OpenBioMaps server.
}
\examples{
#get data rows from the main table from 39980 to 39988
\dontrun{
data <- obm_get('get_data',condition=list(obm_id = '39980:39988'))
}

#get rows from the main table where column 'species' is 'Parus palustris'
\dontrun{
data <- obm_get('get_data',condition=list(species = 'Parus palustris'))
}

#get 100 rows only from filtered query
\dontrun{
data <- obm_get('get_data','limit=100:0',condition=list(species = 'Parus palustris'))
}

#get all data from the default/main table
\dontrun{
data <- obm_get('get_data','*')
}

#get data from a non-default table
\dontrun{
obm_get('get_data','*',table='additional_data')
}

#get list of available forms
\dontrun{
data <- obm_get('get_form_list')
}

#get data of a form
\dontrun{
data <- obm_get('get_form_data',73)
}

#perform strored query 'last' is a custom label
\dontrun{
obm_get('get_report','last')
}

#get list of available tables in the project
\dontrun{
obm_get('get_tables')
}
}
\keyword{get}
