% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{OBM_get}
\alias{OBM_get}
\title{Get Function}
\usage{
OBM_get(scope = "", condition = "", token = OBM$token,
  url = OBM$pds_url, table = OBM$project)
}
\arguments{
\item{scope}{Which scope? e.g. get_data}

\item{condition}{A text condition based on column in your table}

\item{token}{OBM_init() provide it}

\item{url}{OBM_init() provide it}

\item{table}{optional table from project}
}
\description{
This function allows you to get data from an OpenBioMaps server.
}
\examples{
data <- OBM_get('get_data','39980:39988')
data <- OBM_get('get_data','faj=Parus palustris')
get list of available forms
data <- OBM_get('get_form_list',0)
get data of a form
data <- OBM_get('get_form_data',73)
perform strored query 'last' is a custom label
OBM_get('get_report','last')
}
\keyword{get}
