% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{obm_get}
\alias{obm_get}
\title{Get Function}
\usage{
obm_get(scope = "", condition = NULL, token = OBM$token,
  url = OBM$pds_url, table = OBM$project)
}
\arguments{
\item{scope}{Which scope? e.g. get_data}

\item{condition}{A text condition based on column in your table}

\item{token}{obm_init() provide it}

\item{url}{obm_init() provide it}

\item{table}{optional table from project}
}
\description{
This function allows you to get data from an OpenBioMaps server.
}
\examples{
get data rows from the main table from 39980 to 39988
data <- obm_get('get_data','39980:39988')
get rows from the main table where column 'species' is 'Parus palustris'
data <- obm_get('get_data','species=Parus palustris')
get all data from the default/main table
data <- obm_get('get_data','*')
get data from a non-default table
obm_get('get_data','*',table='additional_data')

get list of available forms
data <- obm_get('get_form_list')
get data of a form
data <- obm_get('get_form_data',73)
perform strored query 'last' is a custom label
obm_get('get_report','last')
get list of available tables in the project
obm_get('get_tables')
}
\keyword{get}
