% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{obm_init}
\alias{obm_init}
\title{Init Function}
\usage{
obm_init(
  project = "",
  url = "openbiomaps.org",
  scope = c(),
  verbose = FALSE,
  api_version = 2.3
)
}
\arguments{
\item{project}{Which project?}

\item{url}{project server domain DEFAULT is openbiomaps.org}

\item{scope}{vector of required scopes. DEFAULT is ok usually}

\item{verbose}{print some messages}

\item{api_version}{API version}
}
\value{
boolean
}
\description{
This function is initiating an OBM connection.
}
\examples{
#connect to a database on the default server (openbiomaps.org)
obm_init(project='dead_animals')

#connect on the local server intance to the butterfly database project
\dontrun{
obm_init('http://localhost/biomaps','butterflies')
}
}
\keyword{init}
