% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{obm_init}
\alias{obm_init}
\title{Init Function}
\usage{
obm_init(project = "", url = "openbiomaps.org", scope = c(),
  verbose = F)
}
\arguments{
\item{project}{Which project?}

\item{url}{project server domain DEFAULT is openbiomaps.org}

\item{verbose}{print some messages}
}
\description{
This function allows you to initiate an OBM connection.
}
\examples{
connect to a database on the default server (openbiomaps.org)
obm_init('dead_animals')
connect on the local server intance to the butterfly database project
obm_init('butterflies','http://localhost/biomaps')
}
\keyword{init}
