% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{OBM_init}
\alias{OBM_init}
\title{Init Function}
\usage{
OBM_init(project = "", url = "openbiomaps.org", scope = c(),
  verbose = F)
}
\arguments{
\item{project}{Which project?}

\item{domain}{project server domain DEFAULT is openbiomaps.org}
}
\description{
This function allows you to initiate an OBM connection.
}
\examples{
OBM_init('dead_animals')
OBM_init('butterflies','http://localhost/biomaps')
}
\keyword{init}
