% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{obm_init}
\alias{obm_init}
\title{Init Function}
\usage{
obm_init(url = "openbiomaps.org", project = "", scope = c(),
  verbose = F)
}
\arguments{
\item{url}{project server domain DEFAULT is openbiomaps.org}

\item{project}{Which project?}

\item{verbose}{print some messages}
}
\description{
This function is initiating an OBM connection.
}
\examples{
connect to a database on the default server (openbiomaps.org)
obm_init(project='dead_animals')
connect on the local server intance to the butterfly database project
obm_init('http://localhost/biomaps','butterflies')
}
\keyword{init}
