% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{OBM_put}
\alias{OBM_put}
\title{Put Function}
\usage{
OBM_put(scope = "", condition = "", csv_file = "", form_id = "",
  token = OBM$token, url = OBM$pds_url, table = OBM$project)
}
\arguments{
\item{scope}{currently put_data supported}

\item{condition}{database column names vector, if missing default is the full list from the form}

\item{csv_file}{a csv file with header row}

\item{form_id}{the form's id}

\item{token}{OBM$token}

\item{url}{OBM$url}

\item{table}{OBM$table}
}
\description{
Put Function
}
\examples{
using own list of columns
  OBM_get('get_form_list',0)
  form <- OBM_get('get_form_data',57)
  columns <- unlist(form[,'column'])
  t <- OBM_put('put_data',columns[1:3],form_id=57,csv_file='./teszt2.csv')

using default columns list:
  t <- OBM_put(scope='put_data',form_id=57,csv_file='./teszt2.csv')

}
