% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{obm_refresh_token}
\alias{obm_refresh_token}
\title{Auth Function}
\usage{
obm_refresh_token(
  token = OBM$token$refresh_token,
  url = OBM$token_url,
  client_id = "R",
  verbose = F
)
}
\arguments{
\item{url}{token url: obm_init() provide it}

\item{client_id, }{default is R}

\item{verbose, }{default is FALSE}

\item{refresh}{token}
}
\description{
This function allows you to refresh your OAuth2 token. It is usually a hidden function
}
\examples{
obm_refresh_token(token)
}
\keyword{refresh}
