% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{OBM_refresh_token}
\alias{OBM_refresh_token}
\title{Auth Function}
\usage{
OBM_refresh_token(token = OBM$token$refresh_token, url = OBM$token_url,
  client_id = "R", verbose = F)
}
\arguments{
\item{token}{url OBM_init() provide it
client_id default is R
verbose}
}
\description{
This function allows you to refresh your OAuth2 token. It is usually a hidden function
}
\examples{
OBM_refresh_token(token)
}
\keyword{refresh}
