% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{obm_repo}
\alias{obm_repo}
\title{Repozitorium Function}
\usage{
obm_repo(
  scope = NULL,
  token = OBM$token,
  pds_url = OBM$pds_url,
  data_table = OBM$project,
  params = NULL
)
}
\arguments{
\item{scope}{get or put}

\item{params}{list which contains parameters for repozitorium}
}
\description{
This function allows put data into a repozitorium.
}
\keyword{repozitorium}
