% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{obm_set}
\alias{obm_set}
\title{Set Function}
\usage{
obm_set(scope = "", condition = "", token = OBM$token, url = OBM$pds_url)
}
\arguments{
\item{scope}{Which scope? e.g. set_join}

\item{condition}{A text condition based on column in your table}

\item{token}{obm_init() provide it}

\item{url}{obm_init() provide it}
}
\description{
Experimental function!
This function allows you to set rules for obm_get.
}
\examples{
#automatically join tables
\dontrun{
data <- obm_set('set_join',c('dead_animals','dead_animals_history'))
}
}
\keyword{set}
