% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{OBM_set}
\alias{OBM_set}
\title{Set Function}
\usage{
OBM_set(scope = "", condition = "", token = OBM$token,
  url = OBM$pds_url)
}
\arguments{
\item{scope}{Which scope? e.g. set_join}

\item{condition}{A text condition based on column in your table}

\item{token}{OBM_init() provide it}

\item{url}{OBM_init() provide it}
}
\description{
This function allows you to set rules for OBM_get.
}
\examples{
automatically join tables
data <- OBM_set('set_join',c('dead_animals','dead_animals_history'))
}
\keyword{set}
