% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{obm_sql_query}
\alias{obm_sql_query}
\title{SQL Interface}
\usage{
obm_sql_query(
  sqlcmd,
  username = "",
  password = "",
  paranoid = T,
  port = 5432,
  database = "gisdata"
)
}
\arguments{
\item{sqlcmd}{username most probably automatically set by create_pg_user module
password most probably automatically set by create_pg_user module
paranoid password promt type
postgres server port, default is 5432
database remote database, default is gisdata}
}
\description{
It is a simple SQL Query interface function
}
\examples{
obm_sql_query("SELECT DATE_PART('day', enddate::timestamp - startdate::timestamp) AS days FROM nestboxes WHERE enddate IS NOT NULL AND startdate IS NOT NULL ORDER BY days")
}
\keyword{postgres}
