% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{repo_summary}
\alias{repo_summary}
\title{repo_summary Function}
\usage{
repo_summary(x = NULL)
}
\arguments{
\item{x}{repo output object}
}
\value{
text
}
\description{
This is repozitorium helper function 
Creating readable labels, notes and summary of repo output
}
\keyword{label}
\keyword{summary}
