% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{summary.obm_class}
\alias{summary.obm_class}
\title{summary class Function}
\usage{
\method{summary}{obm_class}(object, ...)
}
\arguments{
\item{object}{obm_class S3 data object}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
object
}
\description{
Create a standard summary for an \code{obm_class} S3 object.
}
\examples{
\dontrun{
summary(obm_data)
}
}
\keyword{summary}
