% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiomaps.R
\name{summary.obm_class}
\alias{summary.obm_class}
\title{summary class Function}
\usage{
\method{summary}{obm_class}(x)
}
\arguments{
\item{obm_class}{S3 data object}
}
\description{
This class function creates a standard summary
}
\examples{
summary(obm.data)
}
\keyword{summary}
